Fixing writeFrame, which could previously experience overflow
In the writeFrame routine, integer types were implicitely retyped to
np.uint32
That triggered the type in seek to be np.uint32 subject to overflow on
large files, the type was replaced by np.uint64.

+++ denpy/DEN.py
@@ -147,9 +147,9 @@ def writeFrame(fileName, k, data, force=False):
 	if len(shape) != 2:
 		raise ValueError('Dimension of data should be 2 %d.' % len(shape))
 	header = np.fromfile(fileName, np.dtype('<i2'), 3)
-	rows = np.uint32(header[0])
-	columns = np.uint32(header[1])
-	k = np.uint32(k)
+	rows = np.uint64(header[0])
+	columns = np.uint64(header[1])
+	k = np.uint64(k)
 	if shape[0] != rows or shape[1] != columns:
 		raise ValueError('There is dimension mismatch between frame (%d, %d) and expected (rows, cols) = (%d, %d) according to header.' %
 						 (rows, columns, shape[0], shape[1]))
